Clean up names
